ghetto defib

+++ code/game/objects/items/trash.dm
@@ -151,9 +151,15 @@
 			new/obj/item/weapon/reagent_containers/food/snacks/customizable/cook/pie(get_turf(src),W)
 			qdel(W)
 			qdel(src)
-	if (istype(W, /obj/item/stack/cable_coil))
+	if (iscablecoil(W))
+		var/obj/item/stack/cable_coil/coil = W
+		if(coil.amount < 5)
+			to_chat(user, "<span class='notice'>There are not enough cables in the stack.</span>")
+			return
+
 		var/obj/item/I = new /obj/item/trash/wired_pietin_assembly(get_turf(src))
-		qdel(W)
+		coil.use(5)
+		to_chat(user, "<span class='notice'>You remove the insulation and wrap the cables around the pie tin.</span>")
 		qdel(src)
 		user.put_in_hands(I)
 
@@ -166,9 +172,10 @@
 
 /obj/item/trash/wired_pietin_assembly/attackby(obj/item/W, mob/user)
 	if (istype(W, /obj/item/trash/pietin))
-		var/obj/item/I = new /obj/item/weapon/melee/defibrillator/ghetto_defibrillator
+		var/obj/item/I = new /obj/item/weapon/melee/defibrillator/improvised
 		qdel(W)
 		qdel(src)
+		to_chat(user, "<span class='notice'>You add a second pie tin to the assembly.</span>")
 		user.put_in_hands(I)
 
 /obj/item/trash/snack_bowl
